pyproject.toml
[tool.poetry]
name = "mestv"
version = "0.1.0"
description = ""
authors = ["pecheritsa <pecheritsa.roman@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
psycopg2 = "^2.9.10"
pillow = "^11.1.0"
python-dotenv = "^1.0.1"
black = "^24.10.0"
isort = "^5.13.2"
django-imagekit = "^5.0.0"
django-multiupload = "^0.6.1"
pytils = "^0.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
